requirements.txt
discord>=1.7.3
mcipc>=2.3.5
PyYAML>=6.0
mcstatus
pystray
pygobject
